Updated CI workflow [logging] middleware now includes cap10 user information
