--- go.mod
@@ -1,9 +1,0 @@
-module github.com/djcass44/go-utils
-
-go 1.15
-
-require (
-	github.com/sirupsen/logrus v1.7.0
-	github.com/stretchr/testify v1.6.1
-	google.golang.org/protobuf v1.26.0
-)
