+++ logging/README.md
@@ -1,1 +1,4 @@
 # Logging
+
+* HTTP Middleware to enrich log messages with OpenTelemetry Trace and Span identifiers
+* HTTP Middleware to enrich log messages with Cap10 user information
