+++ logging/go.mod
@@ -4,11 +4,23 @@ go 1.18
 
 require (
 	github.com/go-logr/logr v1.2.3
+	gitlab.com/av1o/cap10 v0.4.0
 	go.opentelemetry.io/otel v1.7.0
 )
 
 require (
-	github.com/davecgh/go-spew v1.1.1 // indirect
+	github.com/cenkalti/backoff/v4 v4.1.0 // indirect
+	github.com/djcass44/go-tracer v0.2.0 // indirect
+	github.com/djcass44/go-utils v0.2.1-0.20210516001550-3d288882b7f4 // indirect
 	github.com/go-logr/stdr v1.2.2 // indirect
+	github.com/google/go-querystring v1.0.0 // indirect
+	github.com/google/uuid v1.2.0 // indirect
+	github.com/gorilla/mux v1.8.0 // indirect
+	github.com/levigross/grequests v0.0.0-20190908174114-253788527a1a // indirect
+	github.com/sirupsen/logrus v1.8.1 // indirect
 	go.opentelemetry.io/otel/trace v1.7.0 // indirect
+	golang.org/x/crypto v0.0.0-20200820211705-5c72a883971a // indirect
+	golang.org/x/net v0.0.0-20200822124328-c89045814202 // indirect
+	golang.org/x/sys v0.0.0-20211019181941-9d821ace8654 // indirect
+	google.golang.org/protobuf v1.26.0 // indirect
 )
