+++ otel/README.md
@@ -1,1 +1,3 @@
 # OpenTelemetry
+
+Helpers for instrumenting an application with OpenTelemetry.
